build without clerk and database
Pages should not be prerendered if they require access to dynamic data sources
such as clerk and db.
