+++ src/app/account/keys/generate-keys.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest'
 import { mockClerkSession, renderWithProviders } from '@/tests/unit.helpers'
-import { GenerateKeys } from '@/app/account/keys/generate-keys'
+import { GenerateKeys } from './generate-keys'
 import { fireEvent, screen, waitFor } from '@testing-library/react'
 import { UseUserReturn } from '@clerk/types'
 import { useUser } from '@clerk/nextjs'
