+++ src/app/account/keys/page.tsx
@@ -1,8 +1,8 @@
-'use server'
-
-import { GenerateKeys } from '@/app/account/keys/generate-keys'
+import { GenerateKeys } from './generate-keys'
 import { redirect } from 'next/navigation'
 import { getMemberUserPublicKeyAction } from '@/server/actions/user-keys.actions'
+
+export const dynamic = 'force-dynamic'
 
 export default async function KeysPage() {
     const publicKey = await getMemberUserPublicKeyAction()
