+++ src/app/account/signin/page.tsx
@@ -2,6 +2,8 @@ import { SignIn } from './signin'
 import { pageStyles } from '@/styles/common'
 import { Container } from '@/styles/generated/jsx'
 
+export const dynamic = 'force-dynamic'
+
 export default function SigninPage() {
     return (
         <div className={pageStyles}>
