+++ src/app/account/signup/page.tsx
@@ -3,6 +3,8 @@ import { pageStyles } from '@/styles/common'
 import { Container } from '@/styles/generated/jsx'
 import { PROD_ENV } from '@/server/config'
 
+export const dynamic = 'force-dynamic'
+
 export default function SignUpPage() {
     // https://openstax.atlassian.net/browse/OTTER-107
     // Temporarily remove signup page on production
