+++ src/app/admin/organization/page.tsx
@@ -2,6 +2,8 @@ import { MembersAdminTable } from './table'
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query'
 import { fetchMembersAction } from '@/server/actions/member.actions'
 
+export const dynamic = 'force-dynamic'
+
 export default async function MembersAdministration() {
     const queryClient = new QueryClient()
     await queryClient.prefetchQuery({
