--- src/app/api/sentry-example-api/route.ts
@@ -1,9 +1,0 @@
-import { NextResponse } from 'next/server'
-
-export const dynamic = 'force-dynamic'
-
-// A faulty API route to test Sentry's error monitoring
-export function GET() {
-    throw new Error('Sentry Example API Route Error')
-    return NextResponse.json({ data: 'Testing Sentry Error...' })
-}
