+++ src/app/layout.tsx
@@ -8,8 +8,6 @@ import 'mantine-datatable/styles.layer.css'
 import '@mantine/dropzone/styles.layer.css'
 
 import { Providers } from './providers'
-import { AppLayout } from '@/components/layout/app-layout'
-import { RequireMFA } from '@/components/require-mfa'
 import { type ReactNode } from 'react'
 
 export async function generateMetadata(): Promise<Metadata> {
@@ -31,10 +29,7 @@ export default async function RootLayout({
         <html lang="en">
             <body>
                 <Providers>
-                    <AppLayout>
-                        {children}
-                        <RequireMFA />
-                    </AppLayout>
+                    {children}
                 </Providers>
             </body>
         </html>
