+++ src/app/member/[memberIdentifier]/study/[studyIdentifier]/review/page.tsx
@@ -6,10 +6,10 @@ import { getMemberFromIdentifierAction } from '@/server/actions/member.actions'
 import { MemberBreadcrumbs } from '@/components/page-breadcrumbs'
 import { getStudyAction } from '@/server/actions/study.actions'
 import React from 'react'
-import { StudyReviewButtons } from '@/app/member/[memberIdentifier]/study/[studyIdentifier]/review/study-review-buttons'
+import { StudyReviewButtons } from './study-review-buttons'
 import { StudyDetails } from '@/components/study/study-details'
 import { StudyCodeDetails } from '@/components/study/study-code-details'
-import { StudyResults } from '@/app/member/[memberIdentifier]/study/[studyIdentifier]/review/study-results'
+import { StudyResults } from './study-results'
 import { jobStatusForJobAction, latestJobForStudyAction } from '@/server/actions/study-job.actions'
 import { getMemberUserFingerprintAction } from '@/server/actions/user-keys.actions'
