+++ src/app/member/[memberIdentifier]/study/[studyIdentifier]/review/study-results.tsx
@@ -7,7 +7,7 @@ import { StudyJob } from '@/schema/study'
 import { notifications } from '@mantine/notifications'
 import { useMutation, useQuery } from '@tanstack/react-query'
 import { ResultsReader } from 'si-encryption/job-results/reader'
-import { JobReviewButtons } from '@/app/member/[memberIdentifier]/study/[studyIdentifier]/review/job-review-buttons'
+import { JobReviewButtons } from './job-review-buttons'
 import Link from 'next/link'
 import { fingerprintKeyData, pemToArrayBuffer, privateKeyFromBuffer } from 'si-encryption/util'
 import { StudyJobStatus } from '@/database/types'
