+++ src/app/page.tsx
@@ -2,6 +2,8 @@ import { Title } from '@mantine/core'
 import { UserNav } from './user-nav'
 import { pageStyles, mainStyles, footerStyles } from '@/styles/common'
 
+export const dynamic = 'force-dynamic'
+
 // TODO Remove this root page?,
 //  or route users based on their roles to correct pages?
 export default function Home() {
