+++ src/app/researcher/study/request/[memberIdentifier]/study-proposal-form.tsx
@@ -18,7 +18,7 @@ import {
 } from '@mantine/core'
 import { FileDoc, FilePdf, FileText, UploadSimple } from '@phosphor-icons/react/dist/ssr'
 import { UseFormReturnType } from '@mantine/form'
-import { StudyProposalFormValues } from '@/app/researcher/study/request/[memberIdentifier]/study-proposal-schema'
+import { StudyProposalFormValues } from './study-proposal-schema'
 
 export const StudyProposalForm: FC<{
     studyProposalForm: UseFormReturnType<StudyProposalFormValues>
