+++ src/components/layout/app-layout.tsx
@@ -19,6 +19,7 @@ import { OrganizationSwitcher, useAuth } from '@clerk/nextjs'
 import { ReactNode } from 'react'
 import { NavbarItems } from '@/components/layout/navbar-items'
 import { usePathname } from 'next/navigation'
+import { ClerkProvider } from '@clerk/nextjs'
 
 type Props = {
     children: ReactNode
@@ -40,29 +41,31 @@ export function AppLayout({ children }: Props) {
     }
 
     return (
-        <AppShell footer={{ height: 60 }} navbar={{ width: 250, breakpoint: 'xs' }} padding="md">
-            <Notifications position="top-right" />
+        <ClerkProvider>
+            <AppShell footer={{ height: 60 }} navbar={{ width: 250, breakpoint: 'xs' }} padding="md">
+                <Notifications position="top-right" />
 
-            <AppShellNavbar bg={theme.colors.purple[8]}>
-                <Stack py="md">
-                    <AppShellSection>
-                        <Link href="/">
-                            <SafeInsightsLogo />
-                        </Link>
-                    </AppShellSection>
-                    <AppShellSection grow>
-                        <NavbarItems />
-                    </AppShellSection>
-                </Stack>
-            </AppShellNavbar>
-            <AppShellMain bg="#F1F3F5">{children}</AppShellMain>
-            <AppShellFooter p="md" bg={theme.colors.purple[9]}>
-                <Group justify="center" c="white">
-                    <Text>© 2025 - SafeInsights</Text>
-                    {/* TODO Temporary for dev mode only? admins? */}
-                    <OrganizationSwitcher afterSelectOrganizationUrl="/" />
-                </Group>
-            </AppShellFooter>
-        </AppShell>
+                <AppShellNavbar bg={theme.colors.purple[8]}>
+                    <Stack py="md">
+                        <AppShellSection>
+                            <Link href="/">
+                                <SafeInsightsLogo />
+                            </Link>
+                        </AppShellSection>
+                        <AppShellSection grow>
+                            <NavbarItems />
+                        </AppShellSection>
+                    </Stack>
+                </AppShellNavbar>
+                <AppShellMain bg="#F1F3F5">{children}</AppShellMain>
+                <AppShellFooter p="md" bg={theme.colors.purple[9]}>
+                    <Group justify="center" c="white">
+                        <Text>© 2025 - SafeInsights</Text>
+                        {/* TODO Temporary for dev mode only? admins? */}
+                        <OrganizationSwitcher afterSelectOrganizationUrl="/" />
+                    </Group>
+                </AppShellFooter>
+            </AppShell>
+        </ClerkProvider>
     )
 }
