+++ src/components/require-reviewer-keys.tsx
@@ -1,0 +1,23 @@
+'use client'
+
+import { useLayoutEffect } from 'react'
+import { usePathname, useRouter } from 'next/navigation'
+import { getReviewerPublicKeyAction } from '@/server/actions/member.actions'
+
+export const RequireReviewerKeys = () => {
+    const pathname = usePathname()
+    const router = useRouter()
+
+    useLayoutEffect(() => {
+        if (!pathname.startsWith('/account/keys')) {
+            return
+        }
+        getReviewerPublicKeyAction().then(key => {
+            if (!key) {
+                router.push('/account/keys')
+            }
+        })
+    }, [pathname, router])
+
+    return null
+}
